requirements.txt
streamlit
PyGithub
pandas
plotly
transformers
mysql-connector-python
tensorflow
torch
torchvision 
torchaudio